fix: 탈퇴버튼 위치 수정
