+++ client/src/components/pages/mypage/student/settings.tsx
@@ -82,26 +82,6 @@ const StudentSettingsPage = () => {
     }
   };
 
-  const handleDeleteAccount = async () => {
-    const confirm = window.confirm('정말로 회원 탈퇴하시겠습니까? 복구할 수 없습니다.');
-    if (!confirm) return;
-    try {
-      const res = await fetch(`${import.meta.env.VITE_API_URL}/users/me`, {
-        method: 'DELETE',
-        headers: {
-          Authorization: `Bearer ${localStorage.getItem('accessToken')}`,
-        },
-      });
-      if (!res.ok) throw new Error('회원 탈퇴 실패');
-      toast.success('회원 탈퇴가 완료되었습니다.');
-      localStorage.removeItem('accessToken');
-      navigate('/');
-    } catch (err) {
-      if (err instanceof Error) toast.error(err.message);
-      else toast.error('알 수 없는 오류가 발생했습니다.');
-    }
-  };
-
   if (!userInfo) return <div>유저 정보를 불러오는 중입니다...</div>;
 
   return (
@@ -139,8 +119,6 @@ const StudentSettingsPage = () => {
       />
 
       <SaveButton onClick={handlePasswordChange}>비밀번호 변경</SaveButton>
-
-      <WithdrawButton onClick={handleDeleteAccount}>회원 탈퇴</WithdrawButton>
     </Container>
   );
 };
@@ -171,16 +149,6 @@ const SaveButton = styled.button`
   margin-top: 20px;
   padding: 10px 16px;
   background-color: ${({ theme }) => theme.colors.purple};
-  color: white;
-  border: none;
-  border-radius: 6px;
-  cursor: pointer;
-`;
-
-const WithdrawButton = styled.button`
-  margin-top: 40px;
-  padding: 12px;
-  background-color: red;
   color: white;
   border: none;
   border-radius: 6px;
